refactor(gateway): extract root config defaults (GW-2 S8)
Move 65 fn default_*() helpers plus impl Default for Config out of
config.rs into config/defaults.rs. config.rs declares mod defaults
privately and imports via 'use self::defaults::*;' so every existing
#[serde(default = "default_foo")] attribute resolves unchanged.
